Update
Added Job experience information.
Included css at the beginning to "reset" all formatting
Added more css to form desired format.
• Side bar sticks to the side.
• Side fills entire left side vertically.
• Added a fitting color to links.

Profile image is now in the sidebar.
Fixed clicking the image to take you to the top of the page and show name.
Changed names of the HTML and CSS files, so the WIP is the homepage.